azure-backend/requirements.txt
# Azure Functions Python Requirements
azure-functions
azure-cosmos
azure-keyvault-secrets
azure-identity
PyJWT[crypto]
cryptography
requests
PyPDF2
json-repair
